move common codes to utils, parallelize utils / lint
